+++ lib/units.dart
@@ -1,4 +1,8 @@
 library units;
 
+/// All numbers are forced to a max representation of exactly 6 significant digits
+///
+/// For example: 12345678910.12345678 will turn into: 12345678910.123457
+
 export 'src/speed.dart';
 export 'src/length.dart';