+++ CHANGELOG.md
@@ -8,3 +8,10 @@ Included right now:
 
 - Speed
 - Length
+
+## 0.0.2
+
+Minor improvements to Speed and Length, also added some tests
+
+Now, all numbers are forced to a max representation of exactly 6 significant digits
+For example: 12345678910.12345678 will turn into: 12345678910.123457